Raging/ios14

+++ OtherSources/CustomIOSKeyboard/iPad/PKCustomKeyboard.mm
@@ -101,7 +101,7 @@
 	UIButton *button = (UIButton *)sender;
     
     
-    NSString *character = [NSString stringWithFormat:@"D%d", button.tag];
+    NSString *character = [NSString stringWithFormat:@"D%ld", (long) button.tag];
     [self.textView insertText:character];
 }
 
@@ -110,7 +110,7 @@
 	UIButton *button = (UIButton *)sender;
     
     
-    NSString *character = [NSString stringWithFormat:@"U%d", button.tag];
+    NSString *character = [NSString stringWithFormat:@"U%ld", (long) button.tag];
     [self.textView insertText:character];
 }
