+++ OtherSources/CustomIOSKeyboard/iPhone:iPod/PMCustomKeyboard.mm
@@ -99,7 +99,7 @@
 	UIButton *button = (UIButton *)sender;
     
     
-    NSString *character = [NSString stringWithFormat:@"D%d", button.tag];
+    NSString *character = [NSString stringWithFormat:@"D%ld", (long) button.tag];
     [self.textView insertText:character];
 }
 
@@ -108,7 +108,7 @@
 	UIButton *button = (UIButton *)sender;
     
     
-    NSString *character = [NSString stringWithFormat:@"U%d", button.tag];
+    NSString *character = [NSString stringWithFormat:@"U%ld", (long) button.tag];
     [self.textView insertText:character];
 }
