+++ uae4all_gp2x_0.7.2a/src/iphone/RingQ.h
@@ -52,7 +52,7 @@ struct tagRingQ {
 	}
 	
 	inline int count() {
-		int count = in - out;
+		int count = (int) (in - out);
 		if (count < 0)
 			count = size + count;
 		return count;
