+++ uae4all_gp2x_0.7.2a/src/native2amiga.cpp
@@ -89,6 +89,6 @@ uaecptr uae_AllocMem (uae_u32 size, uae_u32 flags)
 void uae_FreeMem (uaecptr memory, uae_u32 size)
 {
     m68k_dreg (regs, 0) = size;
-    m68k_areg (regs, 1) = memory;
+    m68k_areg (regs, 1) = (uint) memory;
     CallLib (get_long (4), -0xD2); /* FreeMem */
 }
